refactor: extract interfaces
