+++ src/types/governance.ts
@@ -1,13 +1,7 @@
-import { DerEncodedPublicKey } from "@dfinity/agent";
-import { Option } from "./option";
-import {
-  AccountIdentifier,
-  CanisterIdString,
-  E8s,
-  NeuronId,
-  PrincipalString,
-} from "./common";
-import { Principal } from "@dfinity/principal";
+import {DerEncodedPublicKey} from '@dfinity/agent';
+import {Option} from './option';
+import {AccountIdentifier, CanisterIdString, E8s, NeuronId, PrincipalString,} from './common';
+import {Principal} from '@dfinity/principal';
 
 export type Action =
   | { RegisterKnownNeuron: KnownNeuron }
@@ -569,4 +563,13 @@ export default interface ServiceInterface {
   claimOrRefreshNeuronFromAccount: (
     request: ClaimOrRefreshNeuronFromAccount
   ) => Promise<NeuronId>;
+}
+
+/**
+ * An error used to ensure at compile-time that it's never reached.
+ */
+export class UnsupportedValueError extends Error {
+  constructor(value: never) {
+    super('Unsupported value: ' + value);
+  }
 }
