+++ src/types/icp_canister.ts
@@ -1,0 +1,4 @@
+export enum FromICPStringError {
+    FRACTIONAL_MORE_THAN_8_DECIMALS,
+    INVALID_FORMAT,
+}
